+++ CHANGELOG.md
@@ -25,8 +25,10 @@ switchTrigger(Trigger.TextSelection);
 
 -   Users can click and drag to highlight an area
 
-**Improvement**
+**Improvements**
 
+-   Keep the expanded/collapsed state of each bookmark
+-   `RenderBookmarkItemProps` includes new `path` property that indicates the path from each bookmark item to the root
 -   `SelectionData` provides more information about the selected text including:
 
 | Property       | Type        | Description                                  |